msvc-drone-back/requirements.txt
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
Flask==3.0.2
Flask-SQLAlchemy==2.5.1
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
idna==3.7
importlib_metadata==7.0.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
marshmallow==3.12.1
marshmallow-sqlalchemy==0.25.0
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pika==1.3.2
pluggy==1.4.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyJWT==2.8.0
pytest==8.1.1
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
rsa==4.9
SQLAlchemy==2.0.28
tomli==2.0.1
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
zipp==3.18.1
